Update Case 2.R
I added the part where we want to determine the lag order of the VAR process. First specify the lag order we want to run (line 60), create numeric vectors to store the IC's (lines 63 and 64). for loop that runs the var for all different horizons and store the corresponding IC's in the numeric vector. choose the lowest IC in lines 75 and 76.